Little headers restructuring
