+++ src/lib/include/black/backends/cvc5
@@ -21,36 +21,9 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#ifndef BLACK_INTERNAL_BACKENDS_CVC5_HPP
-#define BLACK_INTERNAL_BACKENDS_CVC5_HPP
+#ifndef BLACK_BACKENDS_CVC5_HPP
+#define BLACK_BACKENDS_CVC5_HPP
 
-#include <black/logic>
-#include <black/solvers>
+#include <black/internal/solvers/cvc5.hpp>
 
-#include <memory>
-
-namespace black::solvers {
-
-  class cvc5_t : public solver::base
-  {
-  public:
-    cvc5_t();
-    
-    virtual ~cvc5_t() override;
-
-    virtual void set(std::string option, std::string value) override;
-    virtual void set(option opt, std::string value) override;
-    virtual pipes::consumer *consumer() override;
-    virtual support::tribool check() override;
-    virtual std::optional<logic::model> model() const override;
-
-  private:
-    struct impl_t;
-    std::shared_ptr<impl_t> _impl;
-  };
-
-  inline constexpr auto cvc5 = make_solver<cvc5_t>;
-
-}
-
-#endif // BLACK_INTERNAL_BACKENDS_CVC5_HPP
+#endif // BLACK_BACKENDS_CVC5_HPP
