+++ src/lib/include/black/pipes
@@ -28,7 +28,5 @@
 #include <black/logic>
 
 #include <black/internal/pipes/pipeline.hpp>
-#include <black/internal/pipes/pipes.hpp>
-#include <black/internal/pipes/solvers.hpp>
 
 #endif // BLACK_PIPES_HPP
