+++ src/lib/src/solvers/cvc5.cpp
@@ -25,6 +25,7 @@
 #include <black/support>
 #include <black/support/private>
 #include <black/logic>
+#include <black/ast/algorithms>
 #include <black/solvers/cvc5>
 
 #include <cvc5/cvc5.h>
